+++ compiler-tests/src/test/data/diagnostic/functioninject/CannotBeScoped.kt
@@ -1,0 +1,7 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+// ENABLE_TOP_LEVEL_FUNCTION_INJECTION
+<!FUNCTION_INJECT_ERROR!>@SingleIn(AppScope::class)<!>
+@Inject
+fun Example() {
+
+}