+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -6,6 +6,7 @@ import org.jetbrains.kotlin.test.directives.model.SimpleDirectivesContainer
 
 object MetroDirectives : SimpleDirectivesContainer() {
   val GENERATE_ASSISTED_FACTORIES by directive("Enable assisted factories generation.")
+  val ENABLE_TOP_LEVEL_FUNCTION_INJECTION by directive("Enable top-level function injection.")
   val DISABLE_TRANSFORM_PROVIDERS_TO_PRIVATE by
     directive("Disables automatic transformation of providers to be private.")
   val PUBLIC_PROVIDER_SEVERITY by