+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroExtensionRegistrarConfigurator.kt
@@ -55,6 +55,8 @@ class MetroExtensionRegistrarConfigurator(testServices: TestServices) :
         generateAssistedFactories =
           MetroDirectives.GENERATE_ASSISTED_FACTORIES in module.directives,
         transformProvidersToPrivate = transformProvidersToPrivate,
+        enableTopLevelFunctionInjection =
+          MetroDirectives.ENABLE_TOP_LEVEL_FUNCTION_INJECTION in module.directives,
         publicProviderSeverity =
           if (transformProvidersToPrivate) {
             MetroOptions.DiagnosticSeverity.NONE