+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/FirMetroErrors.kt
@@ -134,6 +134,8 @@ internal object FirMetroErrors : BaseDiagnosticRendererFactory() {
   val MEMBERS_INJECT_WARNING by warning1<String>(NAME_IDENTIFIER)
   val MEMBERS_INJECT_RETURN_TYPE_WARNING by warning1<String>(DECLARATION_RETURN_TYPE)
   val DAGGER_REUSABLE_ERROR by error0(NAME_IDENTIFIER)
+  val FUNCTION_INJECT_ERROR by error1<String>(NAME_IDENTIFIER)
+  val FUNCTION_INJECT_TYPE_PARAMETERS_ERROR by error1<String>(TYPE_PARAMETERS_LIST)
 
   override val MAP: KtDiagnosticFactoryToRendererMap =
     KtDiagnosticFactoryToRendererMap("Metro").apply {
@@ -197,6 +199,8 @@ internal object FirMetroErrors : BaseDiagnosticRendererFactory() {
       put(PROVIDES_OR_BINDS_SHOULD_BE_PRIVATE_ERROR, "{0}", STRING)
       put(PROVIDES_OR_BINDS_SHOULD_BE_PRIVATE_WARNING, "{0}", STRING)
       put(PROVIDES_PROPERTIES_CANNOT_BE_PRIVATE, "{0}", STRING)
+      put(FUNCTION_INJECT_ERROR, "{0}", STRING)
+      put(FUNCTION_INJECT_TYPE_PARAMETERS_ERROR, "{0}", STRING)
       put(
         PROVIDER_OVERRIDES,
         "Do not override `@Provides` declarations. Consider using `@ContributesTo.replaces`, `@ContributesBinding.replaces`, and `@DependencyGraph.excludes` instead.",